change pagination, css, integration img in posts

+++ view/backend/adminView.php
@@ -4,17 +4,17 @@
 
 <section id="adminPanel">
 	<div class="jumbotron">
-		<h1 class="display-3 text-center">
+		<h2 class="text-center">
 			<font style="vertical-align: inherit;"> Panel d'administration </font>
-		</h1>
+		</h2>
 	</div>
-	<div class="headPost mb-4">
+	<div class="headPost mb-4 text-center">
 		<h3>Ecrire un article</h3>
 		<button class="btn btn-primary" id="writePost"><a class=" text-white" href="index.php?action=createPost">Ecrire
 				un article</a></button>
 	</div>
 
-	<div class="headPost">
+	<div class="headPost text-center">
 		<h3>Gestion des Articles</h3>
 		<?php
 				if (isset($_GET['update-status']) && $_GET['update-status'] == 'success') {
@@ -32,21 +32,20 @@
 				{
 					if (!empty($post)) {
 			?>
-		<ul class="list-group">
-			<li class="list-group-item d-flex justify-content-between align-items-center">
 
-				<p><a class="linkAdmin"
-						href="index.php?action=updatePost&amp;id=<?= $post['id']; ?>"><?= $post['title']; ?></a></p>
-				<p><?= $post['creation_date_fr']; ?></p>
-				<div id="postModal<?= $countPost++ ?>" class="modal">
-			</li>
-		</ul>
-		<div class="modalContent mb-4">
-		<a class="confirmDelete" href="index.php?action=deletePost&amp;id=<?= $post['id']; ?>">Supprimer l'article <em><?= $post['title']; ?></a>
-			
+			<ul class="list-group">
+				<li class="list-group-item d-flex justify-content-between align-items-center">
+
+					<p><a class="linkAdmin"
+							href="index.php?action=updatePost&amp;id=<?= $post['id']; ?>"><?= $post['title']; ?></a></p>
+					<p><?= $post['date_fr']; ?></p>
+					<div id="postModal<?= $countPost++ ?>" class="modal">
+				</li>
+			</ul>
+		<div class="modalContent">
+			<a class="confirmDelete" href="index.php?action=deletePost&amp;id=<?= $post['id']; ?>">Supprimer l'article <em><?= $post['title']; ?></a>	
 		</div>
-		<a class="report" href="index.php?action=updatePost&amp;id=<?= $post['id']; ?>"></a><p><em><?= $data['date_fr']; ?></em></p>
-
+			<a class="report" href="index.php?action=updatePost&amp;id=<?= $post['id']; ?>"></a><p class="mb-8"><em><?= $post['date_fr']; ?></em></p>
 
 
 		<?php if ($post['date_fr'] < $post['update_date_fr']) {
@@ -57,8 +56,27 @@
 							echo "<p>Pas d'articles !</p>";
 								}
 						}
-					?>
-			</div>
+						$posts->closeCursor();
+
+	if ($nbPage >= 2) {
+?>
+		<section id="Pagination" class="text-right ">
+			<div id="pageFrame">
+<?php
+		for ($i = 1; $i <= $nbPage; $i++) {
+			if ((!isset($_GET['page']) && $i == 1) || (isset($_GET['page']) && $_GET['page'] == $i)) {
+				echo "<span class='cPageFrame'>$i</span>";
+			} else {
+				echo "<a class='pageBlock' href=\"index.php?action=admin&amp;page=$i\">$i</a>";
+			}
+		}
+?>
+		</div>
+<?php
+	}
+
+?>
+		</section>
 
 
 	<div class="headPost mb-4" id="commentManage">
@@ -74,17 +92,27 @@
 		<div class="listPanel">
 			<p><a class="linkAdmin" href="#"><?= $report['author']; ?></a></p>
 			<p><?= $report['comment']; ?></p>
-			<button class="report removeComment"><i class="fas fa-trash-alt"></i></button>
-			<div id="reportModal" class="modal">
-				<div class="modalContent">
-					<p>Voulez-vous vraiment supprimer le commentaire de <em><?= $report['author']; ?></em> ?</p>
-					<a class="confirmDelete"
-						href="index.php?action=deleteComment&amp;id=<?= $report['id']; ?>">Oui</a>
-					<span id="closeCommentModal" class="closeModal">Non</span>
-				</div>
+			<button type="button" class="btn btn-primary report removeComment">Supprimer</button>
+			<div class="modal">
+  				<div class="modal-dialog" role="document">
+    				<div class="modal-content">
+      					<div class="modal-header">
+        					<h5 class="modal-title">Suppression</h5>
+        						<button type="button" class="close" data-dismiss="modal" aria-label="Close">
+          							<span aria-hidden="true">&times;</span>
+        						</button>
+      					</div>
+      					<div class="modal-body">
+        					<p>Voulez-vous vraiment supprimer le commentaire de <em><?= $report['author']; ?></p>
+      					</div>
+      					<div class="modal-footer">
+        					<button type="button" class="btn btn-primary">
+								<a class="confirmDelete" href="index.php?action=deleteComment&amp;id=<?= $report['id']; ?>">Oui</a></button>
+        					<button type="button" class="btn btn-secondary" data-dismiss="modal">Non</button>
+      					</div>
+    				</div>
+  				</div>
 			</div>
-		</div>
-
 		<?php
 	}
 	$reports->closeCursor();
@@ -101,12 +129,12 @@
 			}
 		?>
 
-		<table>
+		<table class="mx-auto">
 			<tr class="table-warning text-center">
-				<td class="w-25 p-3">Identifiant</td>
-				<td class="w-25 p-3">Pseudo</td>
-				<td class="w-25 p-3">Date d'inscription</td>
-				<td class="w-25 p-3">Supprimer</td>
+				<td class="px-3">Id</td>
+				<td class="px-2">Pseudo</td>
+				<td class="px-2">Date <br>d'inscription</td>
+				<td class="px-2">Supprimer</td>
 			</tr>
 			
 			<?php
@@ -115,10 +143,10 @@
 		if(!empty($member)) {
 ?>
 			<tr class="text-center">
-				<td class="w-25 p-3"><?= $member['id']; ?></td>
-				<td class="w-25 p-3"><?= $member['pseudo']; ?></td>
-				<td class="w-25 p-3"><em><?= $member['date_sub']; ?></em></td>
-				<td class="w-25 p-3"><button type="button" class="btn btn-warning">OUI</button>
+				<td><?= $member['id']; ?></td>
+				<td><?= $member['pseudo']; ?></td>
+				<td><em><?= $member['date_sub']; ?></em></td>
+				<td><button type="button" class="btn btn-warning px-1">OUI</button>
 				
 					<div id="memberModal<?= $countMember ?>" class="modal">
 						<div class="modalContent">
